add: first version of script for 2d
base function of ksi and eta, matrix for finite element as well, the matrix of the system (calculating using jacobi coefficients) (without boundary conditions). base functions correct only for first-degree approximation. assuming that the area is rectangular with equidistant nodes. for now system gives the matrix of zeros, although the finite element matrix seems to be correct